🚚 Move instrumented test classes to better match main source set package structure
