+++ app/src/androidTest/java/com/suvanl/fixmylinks/ui/screens/SelectRuleTypeScreenTest.kt
@@ -1,4 +1,4 @@
-package com.suvanl.fixmylinks
+package com.suvanl.fixmylinks.ui.screens
 
 import androidx.activity.ComponentActivity
 import androidx.compose.runtime.Composable
@@ -8,6 +8,7 @@ import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.suvanl.fixmylinks.R
 import com.suvanl.fixmylinks.ui.screens.newruleflow.SelectRuleTypeScreen
 import org.junit.Before
 import org.junit.Rule
